[ManyToMany] Add posibility to add invite human to event

+++ src/main/java/ua/kpi/iasa/IASA_Organiser/service/EventService.java
@@ -4,8 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import ua.kpi.iasa.IASA_Organiser.model.Event;
-import ua.kpi.iasa.IASA_Organiser.model.Tag;
+import ua.kpi.iasa.IASA_Organiser.model.Human;
 import ua.kpi.iasa.IASA_Organiser.repository.EventRepository;
+import ua.kpi.iasa.IASA_Organiser.repository.HumanRepository;
 
 import java.util.Collections;
 import java.util.Comparator;
@@ -22,11 +23,14 @@ public class EventService {
 
     private final EventRepository eventRepository;
 
+    private final HumanRepository humanRepository;
+
     private static final Logger logger = LoggerFactory.getLogger(EventService.class);
 
-    public EventService(CalendarService calendarService, EventRepository eventRepository) {
+    public EventService(CalendarService calendarService, EventRepository eventRepository, HumanRepository humanRepository) {
         this.calendarService = calendarService;
         this.eventRepository = eventRepository;
+        this.humanRepository = humanRepository;
     }
 
     public void createEvent(Event event) {
@@ -66,7 +70,24 @@ public class EventService {
     }
 
     public void deleteById(UUID id) {
-        eventRepository.deleteById(id);
+        Optional<Event> optEvent = eventRepository.findById(id);
+        if (optEvent.isEmpty()) {
+            return;
+        }
+        Event event = optEvent.get();
+        if (event.getInvited().isEmpty()) {
+            eventRepository.deleteById(id);
+            return;
+        }
+        List<Human> allHumans = humanRepository.findAllById(event.getInvited().stream().map(Human::getId).collect(Collectors.toSet()));
+        allHumans.forEach(human -> {
+            Set<Event> events = human.getEvents().stream().filter(event1 -> !event1.getId().equals(event.getId())).collect(Collectors.toSet());
+            human.setEvents(events);
+            humanRepository.save(human);
+        });
+        event.setInvited(Collections.emptySet());
+        eventRepository.save(event);
+        eventRepository.delete(event);
     }
 
     List<Event> sortEventsByPriority(List<Event> events) {
@@ -81,14 +102,6 @@ public class EventService {
 
     public void deleteListEvent(List<Event> events) {
         eventRepository.deleteAll(events);
-    }
-
-    public Set<Tag> getEventTags(UUID eventId) {
-        Set<Tag> eventTags = eventRepository.findById(eventId).get().getTags();
-        if (eventTags == null) {
-            return Collections.emptySet();
-        }
-        return eventTags;
     }
 
 }
